+++ runway/cfngin/commands/stacker/__init__.py
@@ -1,8 +1,9 @@
 """CFNgin commands."""
 import logging
+import sys
 import warnings
 
-from runway.cfngin import __version__
+from runway import cfngin, variables
 
 from ... import session_cache
 from ...config import render_parse_load as load_config
@@ -16,6 +17,10 @@ from .graph import Graph
 from .info import Info
 
 LOGGER = logging.getLogger(__name__)
+
+# TODO remove - https://github.com/onicagroup/runway/issues/293
+sys.modules["stacker"] = cfngin  # shim to remove stacker dependency
+sys.modules["stacker.variables"] = variables  # shim to support standard variables
 
 
 class Stacker(BaseCommand):
@@ -67,5 +72,7 @@ class Stacker(BaseCommand):
     def add_arguments(self, parser):
         """Add CLI arguments."""
         parser.add_argument(
-            "--version", action="version", version="%%(prog)s %s" % (__version__,)
+            "--version",
+            action="version",
+            version="%%(prog)s %s" % (cfngin.__version__,),
         )