+++ runway/core/providers/aws/__init__.py
@@ -1,6 +1,16 @@
 """Runway AWS objects."""
+from . import s3
 from ._account import AccountDetails
 from ._assume_role import AssumeRole
 from ._cli import cli
+from ._response import BaseResponse, ResponseError, ResponseMetadata
 
-__all__ = ["AccountDetails", "AssumeRole", "cli"]
+__all__ = [
+    "AccountDetails",
+    "AssumeRole",
+    "BaseResponse",
+    "ResponseError",
+    "ResponseMetadata",
+    "cli",
+    "s3",
+]